+++ CHANGELOG.md
@@ -100,6 +100,7 @@
 * Instance 3rd pane: Add consortial holdings/item accordion. Refs UIIN-2410.
 * Consortial Central Tenant: Handling Holdings and Item actions on the Instance detail view. Refs UIIN-2523.
 * Hide Held by facet in Inventory contributor and subject browse. Refs UIIN-2591.
+* Enable/disable consortial holdings/item actions based on User permissions. Refs UIIN-2452.
 
 ## [9.4.12](https://github.com/folio-org/ui-inventory/tree/v9.4.12) (2023-09-21)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v9.4.11...v9.4.12)